+++ src/memory/mbc/mbc2.hpp
@@ -48,12 +48,12 @@ class MBC2 : public MemoryBankController
     /**
      * The size of a ROM bank in bytes
      */
-    static const int ROM_BANK_SIZE_IN_BYTES = 16_KiB;
+    static const size_t ROM_BANK_SIZE_IN_BYTES = 16_KiB;
 
     /**
      * The size of the RAM in bytes
      */
-    static const int RAM_SIZE_IN_BYTES = 512_KiB;
+    static const size_t RAM_SIZE_IN_BYTES = 512_KiB;
 
     /**
      * The id of the ROM bank currently selected.
